change FPGA kernel pointer type

+++ src/operators/kernel/fpga/conv_add_bn_relu_kernel.cpp
@@ -57,7 +57,7 @@ bool ConvAddBNReluKernel<FPGA, float>::Init(FusionConvAddBNReluParam *param) {
   param->SetNewScale(new_scale);
   param->SetNewBias(new_bias);
   fpga::quantify_filter(filter);
-  auto filter_ptr = filter->data<float>();
+  auto filter_ptr = filter->data<int8_t>();
 
   fpga::ConvArgs convArgs;
   convArgs.relu_enabled = relu_enabled;
